AML/requirements.txt
pandas==2.2.3
numpy==2.1.3
pymongo==4.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
PyMySQL==1.1.1
mysql-connector-python==9.0.0
neo4j==5.24.0
scikit-learn==1.5.2
web3==6.19.0
networkx>=3.2
